requirements.txt
django==4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.0
Pillow==10.0.0
django-filter==23.3
python-decouple==3.8
twilio==8.10.0
groq==0.4.2
numpy==1.26.4
opencv-python==4.10.0.84
ultralytics==8.3.100
